🌃 renamed component from 'Blog' to 'TextCard'
