+++ src/routes/index.svelte
@@ -4,7 +4,7 @@
 	import Hero from '../components/Hero.svelte';
 	import CTA from '../components/CTA.svelte';
 	import SectionHeader from '../components/SectionHeader.svelte';
-	import Blog from '../components/Blog.svelte';
+	import TextCard from '../components/TextCard.svelte';
 	import { blog } from '../content/BlogJSON.svelte';
 	import { article } from '../content/ArticleJSON.svelte';
 	import { estimateTimeToRead, truncateText } from '../components/Functions.svelte';
@@ -19,7 +19,7 @@
 <SectionHeader text="Read Blogs" textColor="text-blue-500" divColor="divide-blue-500" />
 <div class="w-5/6 mx-auto grid md:grid-cols-3 gap-4">
 	{#each blog as blog}
-		<Blog
+		<TextCard
 			title={blog.title}
 			subtitle={truncateText(blog.content)}
 			img={blog.img}
@@ -32,7 +32,7 @@
 <SectionHeader text="Read Articles" textColor="text-yellow-500" divColor="divide-yellow-500" />
 <div class="w-5/6 mx-auto grid md:grid-cols-3 gap-4">
 	{#each article as article}
-		<Blog
+		<TextCard
 			title={article.title}
 			subtitle={truncateText(article.content)}
 			img={article.img}
